+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased
 - GCGI-1109: Update "Patient Genetic Sex" to "Patient sex assigned at birth"
 - GCGI-1544: Handle missing or empty HLA data
+- GCGI-1539: Environment variable for CouchDB credentials
 - GCGI-1546: Fix installation of INI templates for benchmark script
 - Added SETD2 to Captiv8 SWISNF genes